agente_csv/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openai
pyzipper
tqdm
streamlit
langchain
tiktoken
langchain-community
